Make dtype explicit (attempts to fix windows failure)

+++ sdks/python/apache_beam/dataframe/frames_test.py
@@ -284,13 +284,15 @@ class DeferredFrameTest(_AbstractFrameTest):
     self._run_test(lambda df: df.sort_index(axis=1, na_position='first'), df)
 
   def test_where_callable_args(self):
-    df = pd.DataFrame(np.arange(10).reshape(-1, 2), columns=['A', 'B'])
+    df = pd.DataFrame(np.arange(10, dtype=np.int64).reshape(-1, 2),
+                      columns=['A', 'B'])
 
     self._run_test(
         lambda df: df.where(lambda df: df % 2 == 0, lambda df: df * 10), df)
 
   def test_where_concrete_args(self):
-    df = pd.DataFrame(np.arange(10).reshape(-1, 2), columns=['A', 'B'])
+    df = pd.DataFrame(np.arange(10, dtype=np.int64).reshape(-1, 2),
+                      columns=['A', 'B'])
 
     self._run_test(
         lambda df: df.where(
